requirements.txt
streamlit
langchain
langchain-text-splitters
langchain-community
langchain-google-genai
faiss-cpu
pypdf
sentence-transformers
python-dotenv